requirements.txt
Pillow==8.3.2
termcolor==1.1.0
